+++ lib/components/box_victory.dart
@@ -23,7 +23,6 @@ class BoxVictory extends PositionComponent
 
   @override
   void onTapUp(TapUpEvent event) {
-    gameRef.NewGame();
     gameRef.router.pop();
   }
 }